authorization for campground and review crud

+++ routes/campgrounds.js
@@ -1,21 +1,12 @@
 const express = require("express");
 const router = express.Router();
 const catchAsync = require("../utils/catchAsync");
-const ExpressError = require("../utils/ExpressError");
 const Campground = require("../models/campground");
-const { campgroundSchema } = require("../schemas.js");
-
-const validateCampground = function (req, res, next) {
-    const { error } = campgroundSchema.validate(req.body);
-    if (error) {
-        const message = error.details
-            .map((element) => element.message)
-            .join(",");
-        throw new ExpressError(message, 400);
-    } else {
-        next();
-    }
-};
+const {
+    isLoggedIn,
+    isAuthorOfCampground,
+    validateCampground,
+} = require("../middleware");
 
 router.get(
     "/",
@@ -25,15 +16,17 @@ router.get(
     })
 );
 
-router.get("/new", function (req, res) {
+router.get("/new", isLoggedIn, function (req, res) {
     res.render("campgrounds/new");
 });
 
 router.post(
     "/",
+    isLoggedIn,
     validateCampground,
     catchAsync(async function (req, res, next) {
         const campground = new Campground(req.body.campground);
+        campground.author = req.user._id;
         await campground.save();
         req.flash("success", "Successfully made a new campground!");
         res.redirect(`/campgrounds/${campground._id}`);
@@ -43,9 +36,14 @@ router.post(
 router.get(
     "/:id",
     catchAsync(async function (req, res) {
-        const campground = await Campground.findById(req.params.id).populate(
-            "reviews"
-        );
+        const campground = await Campground.findById(req.params.id)
+            .populate({
+                path: "reviews",
+                populate: {
+                    path: "author",
+                },
+            })
+            .populate("author");
         if (!campground) {
             req.flash("error", "Campground not found!");
             return res.redirect("/campgrounds");
@@ -57,6 +55,8 @@ router.get(
 //update by id
 router.get(
     "/:id/edit",
+    isLoggedIn,
+    isAuthorOfCampground,
     catchAsync(async function (req, res) {
         const campground = await Campground.findById(req.params.id);
         if (!campground) {
@@ -69,6 +69,8 @@ router.get(
 
 router.put(
     "/:id",
+    isLoggedIn,
+    isAuthorOfCampground,
     validateCampground,
     catchAsync(async function (req, res) {
         const campground = await Campground.findByIdAndUpdate(req.params.id, {
@@ -82,7 +84,14 @@ router.put(
 //delete by id
 router.delete(
     "/:id",
+    isLoggedIn,
+    isAuthorOfCampground,
     catchAsync(async function (req, res) {
+        const campground = await Campground.findById(req.params.id);
+        if (!campground) {
+            req.flash("error", "Campground not found!");
+            return res.redirect("/campgrounds");
+        }
         await Campground.findByIdAndDelete(req.params.id);
         req.flash("success", "Successfully removed a campground!");
         res.redirect("/campgrounds");
